+++ docs/changes.rst
@@ -4,7 +4,8 @@ Changelog
 10.0.0b2 (unreleased)
 ---------------------
 
-- Nothing changed yet.
+- A new view ``@@refresh-resources-timestamp`` has been introduced
+  to break the browser cache
 
 
 10.0.0b1 (2018-10-10)